+++ src/org/parabuild/ci/webui/admin/ResumeGroupPanel.java
@@ -166,7 +166,7 @@ public final class ResumeGroupPanel extends MessagePanel {
     final BuildListService buildListService = ServiceManager.getInstance().getBuildListService();
     for (int i = 0; i < selectedBuildIDs.size(); i++) {
       final Integer buildID = (Integer) selectedBuildIDs.get(i);
-      final BuildService buildService = buildListService.getBuild(buildID.intValue());
+      final BuildService buildService = buildListService.getBuild(buildID);
       if (buildService == null) {
         continue;
       }
@@ -300,7 +300,7 @@ public final class ResumeGroupPanel extends MessagePanel {
     final BuildListService buildListService = ServiceManager.getInstance().getBuildListService();
     for (final Iterator iterator = result.iterator(); iterator.hasNext();) {
       final Integer buildID = (Integer) iterator.next();
-      buildListService.getBuild(buildID.intValue()).resumeBuild();
+      buildListService.getBuild(buildID).resumeBuild();
     }
 
     return createDoneResult();
